Modified load screen for the card and added method for suspend card button by checking all the scenarios.

+++ phase2/src/hotcupsofjava/transitsystemmanager/objects/userobjects/Card.java
@@ -65,7 +65,7 @@ public class Card extends TransitSystemObject {
         return suspended;
     }
 
-    public void addBalance(int amount) {
+    public void addBalance(double amount) {
         balance += amount;
     }
